requirements.txt
psutil==5.9.5
python-dotenv==1.0.1
plyer==2.1.0
win10toast==0.9